Move: 해시태그 컨테이너 박스 공통 컴포넌트로 분리
